pyproject.toml
[project]
name = "image-creator-mcp"
version = "0.1.0"
description = "AI-first MCP server that turns image APIs into one stable file-output tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp[cli]>=1.2.0",
  "httpx>=0.27,<1",
  "python-dotenv>=1,<2",
]

[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_creator"]

[project.scripts]
image-creator-mcp = "image_creator.server:main"

[dependency-groups]
dev = [
  "pytest>=8,<9",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
